IBM8514A: Fix Bresenham to S3 Hardware Correct - fix line drag in X11 Tru64
This fixes line dragging around windows in X11 Tru64 (OpenVMS uses vector drawing instead....)
Also fixed a potential truncation bug

+++ src/ibm8514a.h
@@ -175,7 +175,7 @@ protected:
 
   required_device<svga_device> m_vga;  // for pass-through
 private:
-  void ibm8514_draw_vector(uint8_t len, uint8_t dir, bool draw);
+  void ibm8514_draw_vector(uint16_t len, uint8_t dir, bool draw);
   void ibm8514_wait_draw_ssv();
   void ibm8514_draw_ssv(uint8_t data);
   void ibm8514_wait_draw_vector();
